+++ lib/main.dart
@@ -1,6 +1,5 @@
+import 'package:biasiswa_yp/src/login/login_page.dart';
 import 'package:flutter/material.dart';
-
-import 'pages/login/login_page.dart';
 
 void main() {
   runApp(MyApp());
