+++ pubspec.yaml
@@ -23,16 +23,12 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-    expandable: ^4.1.4
+  expandable: ^4.1.4
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
